LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGGraphicsElement.getBBox-03-expected.txt


FAIL rect1 assert_equals: rect1: x expected 1 but got 0
FAIL rect2 assert_equals: rect2: x expected 1 but got 0
FAIL circle assert_equals: circle: x expected 1 but got 0
PASS ellipse1
PASS ellipse2
PASS image3
PASS image4
FAIL foreign1 assert_equals: foreign1: x expected 1 but got 0
FAIL foreign2 assert_equals: foreign2: x expected 1 but got 0
PASS SVGGraphicsElement.prototype.getBBox for elements with negative sizes
